requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.3
django-filter==24.1
djangorestframework==3.14.0
filelock==3.13.1
imageio==2.34.0
lazy_loader==0.3
Markdown==3.5.2
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
platformdirs==4.1.0
pytz==2023.3.post1
scikit-image==0.22.0
scipy==1.12.0
sqlparse==0.4.4
tifffile==2024.2.12
tzdata==2023.4
virtualenv==20.25.0
